json type matching int->unsigned long (#3)

+++ middleware/MA/MA.c
@@ -205,7 +205,7 @@ void MQTTMessageArrived(char* topic, char* msg, int msgLen) {
                 ArrayElement* resultArray = calloc(1, sizeof(ArrayElement));
                 resultArray->capacity = 2;
                 resultArray->element = calloc(1, sizeof(Element) * resultArray->capacity);
-                int errorCode = 106;
+                unsigned long errorCode = 106;
 
                 Element* item = resultArray->element + resultArray->total;
                 item->type = JSON_TYPE_LONG;
@@ -251,7 +251,7 @@ void MQTTMessageArrived(char* topic, char* msg, int msgLen) {
             if(!attribute) return;
             cJSON* act7colorLedObject = cJSON_GetObjectItemCaseSensitive(attribute, "act7colorLed");
             if(!act7colorLedObject) return;
-            int act7colorLed = act7colorLedObject->valueint;
+            unsigned long act7colorLed = act7colorLedObject->valueint;
             SKTDebugPrint(LOG_LEVEL_INFO, "act7colorLed : %d, %d", act7colorLed, cmdId);
             int rc = RGB_LEDControl(act7colorLed);
             if(rc != 0) {
@@ -447,7 +447,7 @@ static void attribute() {
     item = arrayElement->element + arrayElement->total;
     item->type = JSON_TYPE_LONG;
     item->name = "sysErrorCode";
-    int errorCode = 0;
+    unsigned long errorCode = 0;
     item->value = &errorCode;
     arrayElement->total++;
 
@@ -487,7 +487,7 @@ static void attribute() {
     item = arrayElement->element + arrayElement->total;
     item->type = JSON_TYPE_LONG;
     item->name = "act7colorLed";
-    int act7colorLed = 0;
+    unsigned long act7colorLed = 0;
     item->value = &act7colorLed;
     arrayElement->total++;
 
@@ -510,7 +510,7 @@ static void attribute() {
     SRAConvertCSVData( csv_attr, "1.0");
     //Serial
     SRAConvertCSVData( csv_attr, "710DJC5I10000290");
-    int errorCode = 0;
+    unsigned long errorCode = 0;
     snprintf( tmp, 64, "%d", errorCode);
     //Error code
     SRAConvertCSVData( csv_attr, tmp);
@@ -528,7 +528,7 @@ static void attribute() {
     //Longitude
     SRAConvertCSVData( csv_attr, "127.115479"); 
     //Led
-    int act7colorLed = 0;
+    unsigned long act7colorLed = 0;
     snprintf( tmp, 64, "%d", act7colorLed );
     SRAConvertCSVData( csv_attr, tmp);
     tpSimpleRawAttribute(csv_attr, FORMAT_CSV);
